requirements.txt
torch==2.10.0
easyocr==1.7.2
torchvision==0.25.0
supervision==0.18.0
openai==1.3.5
transformers==4.49.0
ultralytics==8.3.70
azure-identity==1.25.1
numpy==1.26.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
gradio==6.5.1
dill==0.4.1
accelerate==1.12.0
timm==1.0.24
einops==0.8.0
paddlepaddle==3.2.2
paddleocr==3.4.0
ruff==0.6.7
pre-commit==3.8.0
pytest==8.3.3
pytest-asyncio==0.23.6
pyautogui==0.9.54
streamlit==1.54.0
anthropic[bedrock,vertex]==0.79.0
jsonschema==4.22.0
boto3==1.42.44
google-auth==2.48.0
screeninfo==0.8.1
uiautomation==2.0.29
dashscope==1.25.11
groq==1.0.0
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.12.5
PyYAML==6.0.2
python-multipart==0.0.22
tokenizers==0.21.4
huggingface_hub==0.36.2
safetensors==0.7.0
sentencepiece==0.2.1
paddlex==3.4.1
